23586: [EDF] HYDRO: Copy mesh to new geometry
--- GEOM_Gen_i.cc ---
In order to assure safe and fast search of initial groups by groups of
the new geometry, behavior of GEOM_Gen operations that implement searching
and publishing of new groups corresponding to initial groups, is
slightly modified. A reference to an initial sub-object is published
under a corresponding new sub-object. The modified functions are
RestoreSubShapesO(), RestoreGivenSubShapesO() and RestoreSubShapesSO().

--- GEOM_Engine.cxx ---
Avoid double removal of GEOM_BaseObject that leads to removal of all
attributes from the root TDF label. That caused some regression, don't
remember which.